+++ tests/profiles/googlefonts_test.py
@@ -2009,9 +2009,8 @@ def test_check_metadata_valid_name_values():
 
     # Good font with other language name entries
     font = TEST_FILE("bizudpmincho-nameonly/BIZUDPMincho-Regular.ttf")
-    ttfont = TTFont(font)
 
-    assert_PASS(check(ttfont), "with a good font with other languages...")
+    assert_PASS(check(font), "with a good font with other languages...")
 
 
 def test_check_metadata_valid_full_name_values():