+++ CHANGELOG.md
@@ -15,7 +15,7 @@ A more detailed list of changes is available in the corresponding milestones for
 
 ### Changes to existing checks
 #### On the Google Fonts Profile
-  - **[com.google.fonts/check/metadata/valid_name_values]:**  Compare METDATA family name against the font's best family name (issue #4262)
+  - **[com.google.fonts/check/metadata/valid_name_values]:**  Compare METADATA family name against the font's best family name (issue #4262)
 
 
 ## 0.10.8 (2023-Dec-15)